requirements.txt
alembic>=1.17.0
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.3
autopep8==1.5.4
bcrypt==3.2.0
blinker>=1.9.0
convertdate==2.3.0
decorator==4.4.2
distlib==0.3.0
Flask>=3.1.2
Flask-Dance>=7.1.0
Flask-JWT-Extended>=4.6.0
Flask-Login>=0.6.3
Flask-Mail>=0.10.0
Flask-Migrate>=4.0.7
Flask-Session==0.3.2
Flask-SQLAlchemy>=3.1.1
idna>=2.10
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous>=2.2.0
jedi==0.17.0
Jinja2>=3.1.6
MarkupSafe>=3.0.3
oauthlib>=3.3.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT>=2.10.1
PyMySQL>=1.1.0
mysql-connector-python>=8.0.0
python-dateutil>=2.9,<3.0
python-editor==1.0.4
requests>=2.25.0
requests-oauthlib>=1.4.0
six==1.14.0
SQLAlchemy>=2.0.44,<2.1
SQLAlchemy-Utils>=0.41.2
sqlparse==0.3.0
urllib3>=1.26.2
URLObject==2.4.3
virtualenv==20.0.20
wcwidth==0.1.9
Werkzeug>=3.1.3
zipp==3.1.0
